pyproject.toml
[project]
name = "amap-shp"
version = "0.2.0"
description = "下载高德地图行政区划数据"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "frykit>=0.7.6",
    "loguru>=0.7.3",
    "prcoords>=1.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "retrying>=1.4.2",
]

[project.scripts]
main = "amap_shp.__main__:main"

[dependency-groups]
dev = [
    "ipython>=9.7.0",
    "ruff>=0.14.3",
]

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "RUF022"]
ignore = ["F403", "E731"]

[tool.basedpyright]
typeCheckingMode = "standard"
reportMissingTypeStubs = false
reportUnnecessaryTypeIgnoreComment = true
